tools/calib_record_tool/pyproject.toml
[project]
name = "stereo-charuco-pipeline"
version = "0.1.0"
description = "Stereo ChArUco 3D motion capture pipeline with auto-calibration"
requires-python = ">=3.10,<3.13"

dependencies = [
    "caliscope>=0.6.9",
    "pyyaml>=6.0",
    "opencv-contrib-python>=4.8.0.74",
    "Pillow>=10.0.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
stereo-pipeline = "recorder.cli:main"
stereo-calibrate = "recorder.cli:calibrate_only"
stereo-record = "recorder.cli:pipeline_only"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/recorder"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.sdist]
exclude = [
    "bin/**",
    "_test/**",
    "**/*.mp4",
    "**/*.avi",
]
